changed font loading to use a memory arena

+++ ElbranEngine2/cpp/MemoryArena.h
@@ -5,15 +5,15 @@ class MemoryArena {
 public:
 	MemoryArena() = default;
 
-	void Allocate(size_t size);
-	void* Reserve(size_t bytes);
-	MemoryArena ReserveSubArena(size_t bytes);
+	void Allocate(uint64_t size);
+	void* Reserve(uint64_t bytes);
+	MemoryArena ReserveSubArena(uint64_t bytes);
 	void Clear();
 	void Release();
 
 private:
 	uint8_t* data;
 	uint8_t* next;
-	size_t size;
+	uint64_t size;
 };
